scripts/scripts/requirements.txt
libsql-client>=0.3.0
python-dotenv>=1.0.0
